requirements.txt
streamlit
yfinance
pandas
pandas_datareader
plotly
python-dotenv
requests
scikit-learn
statsmodels
fredapi
flask
